Commit patch from Bjorn Persson. Store SPI values as unsigned int and make port numbers more consistent.

+++ ser_ims/trunk/modules/pcscf/registrar_storage.c
@@ -370,7 +370,7 @@ void free_r_public(r_public *p)
  * @param mod - The IPSec mode (either 'transport' or 'tunnel')
  * @returns the new r_ipsec* container or NULL on error
  */
-r_ipsec* new_r_ipsec(int spi_uc,int spi_us,int spi_pc,int spi_ps,int port_uc,int port_us,
+r_ipsec* new_r_ipsec(unsigned int spi_uc,unsigned int spi_us,unsigned int spi_pc,unsigned int spi_ps,unsigned short int port_uc,unsigned short int port_us,
 	str ealg_setkey,str r_ealg, str ck_esp,str alg_setkey,str r_alg, str ik_esp, str prot, str mod)
 {
 	r_ipsec *ipsec;
